+++ objc2-foundation/src/macros.rs
@@ -6,43 +6,79 @@
 /// and it's instances must have the raw encoding `Encoding::Object` (an
 /// example: `NSAutoreleasePool` does not have this). Finally the ownership
 /// must be correct for this class.
-macro_rules! object_struct {
-    (unsafe $name:ident) => {
+macro_rules! object {
+    (
+        $(#[$m:meta])*
+        unsafe $v:vis struct $name:ident
+    ) => {
+        object!($(#[$m])* unsafe $v struct $name<> {});
+    };
+    (
+        $(#[$m:meta])*
+        unsafe $v:vis struct $name:ident<$($t:ident $(: $b:ident)?),*> {
+            $($p:ident: $pty:ty,)*
+        }
+    ) => {
         // TODO: `extern type`
+        $(#[$m])*
         #[repr(C)]
-        pub struct $name {
+        $v struct $name<$($t $(: $b)?),*> {
             _private: [u8; 0],
+            $($p: $pty),*
         }
 
-        unsafe impl ::objc2::Message for $name {}
+        unsafe impl<$($t $(: $b)?),*> ::objc2::Message for $name<$($t),*> { }
 
-        unsafe impl ::objc2::RefEncode for $name {
+        unsafe impl<$($t $(: $b)?),*> ::objc2::RefEncode for $name<$($t),*> {
             const ENCODING_REF: ::objc2::Encoding<'static> = ::objc2::Encoding::Object;
         }
 
-        unsafe impl $crate::INSObject for $name {
+        unsafe impl<$($t $(: $b)?),*> $crate::INSObject for $name<$($t),*> {
             fn class() -> &'static ::objc2::runtime::Class {
                 ::objc2::class!($name)
             }
         }
 
-        impl ::core::cmp::PartialEq for $name {
+        // Objective-C equality has approximately the same semantics as Rust
+        // equality (although less aptly specified).
+        //
+        // At the very least, equality is _expected_ to be symmetric and
+        // transitive, and that's about the best we can do.
+        //
+        // `T: PartialEq` bound added because e.g. `NSArray` does deep
+        // (instead of shallow) equality comparisons.
+        //
+        // See also https://nshipster.com/equality/
+        impl<$($t: ::core::cmp::PartialEq $(+ $b)?),*> ::core::cmp::PartialEq for $name<$($t),*> {
+            #[inline]
             fn eq(&self, other: &Self) -> bool {
                 use $crate::INSObject;
                 self.is_equal(other)
             }
         }
 
-        impl ::core::cmp::Eq for $name {}
+        // Most types' equality is reflexive.
+        //
+        // `T: Eq` bound added to prevent e.g. `NSValue<f32>` from being `Eq`
+        // (even though `[NAN isEqual: NAN]` is true in Objective-C).
+        impl<$($t: ::core::cmp::Eq $(+ $b)?),*> ::core::cmp::Eq for $name<$($t),*> {}
 
-        impl ::core::hash::Hash for $name {
+        // Hashing in Objective-C has the exact same requirement as in Rust:
+        //
+        // > If two objects are equal (as determined by the isEqual: method),
+        // > they must have the same hash value.
+        //
+        // See https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418859-hash
+        impl<$($t: ::core::hash::Hash $(+ $b)?),*> ::core::hash::Hash for $name<$($t),*> {
+            #[inline]
             fn hash<H: ::core::hash::Hasher>(&self, state: &mut H) {
                 use $crate::INSObject;
                 self.hash_code().hash(state);
             }
         }
 
-        impl ::core::fmt::Debug for $name {
+        // TODO: Consider T: Debug bound
+        impl<$($t $(: $b)?),*> ::core::fmt::Debug for $name<$($t),*> {
             fn fmt(&self, f: &mut ::core::fmt::Formatter<'_>) -> ::core::fmt::Result {
                 use $crate::{INSObject, INSString};
                 ::objc2::rc::autoreleasepool(|pool| {
@@ -51,27 +87,6 @@ macro_rules! object_struct {
             }
         }
     };
-}
-
-/// TODO
-///
-/// # Safety
-///
-/// The given type must be valid as an Objective-C object. TODO: More.
-macro_rules! object_impl {
-    (unsafe $name:ident) => (
-        object_impl!(unsafe $name,);
-    );
-    (unsafe $name:ident<$($t:ident$(: $b:ident)?),+>) => (
-        object_impl!(unsafe $name, $($t$(: $b)?),+);
-    );
-    (unsafe $name:ident, $($t:ident$(: $b:ident)?),*) => (
-        unsafe impl<$($t$(:($b))?),*> ::objc2::Message for $name<$($t),*> { }
-
-        unsafe impl<$($t$(: $b)?),*> ::objc2::RefEncode for $name<$($t),*> {
-            const ENCODING_REF: ::objc2::Encoding<'static> = ::objc2::Encoding::Object;
-        }
-    );
 }
 
 macro_rules! unsafe_def_fn {